Fix windows time management

+++ include/timeman.h
@@ -31,8 +31,9 @@ extern volatile atomic_int uciponder;
 /* timepoint_t is given in nanoseconds. */
 typedef int64_t timepoint_t;
 
-#define TPPERSEC 1000000000l
-#define TPPERMS     1000000l
+/* Cast to fix linux/windows discrepancy. */
+#define TPPERSEC ((int64_t)1000000000ll)
+#define TPPERMS  ((int64_t)1000000ll)
 
 struct timeinfo {
 	int stop_on_time;
